feat: remove paw.ts, remove CustomClient
