+++ src/commands/miscellaneous/help.ts
@@ -1,6 +1,7 @@
-import { ActionRowBuilder, APIEmbedField, EmbedBuilder, SelectMenuBuilder, SelectMenuInteraction, SlashCommandBuilder } from 'discord.js';
+import { ActionRowBuilder, APIEmbedField, Client, EmbedBuilder, SelectMenuBuilder, SelectMenuInteraction, SlashCommandBuilder } from 'discord.js';
 import { getArrayElement, respond, update } from '../../utils/helperFunctions';
-import { CustomClient, SlashCommand } from '../../typedef';
+import { SlashCommand } from '../../typedef';
+import { handle } from '../..';
 
 const { default_color, maksi, ezra, ren, jags, elliott, hazenith, johanna, sky, asriel } = require('../../../config.json');
 const { version } = require('../../../package.json');
@@ -38,7 +39,7 @@ export const command: SlashCommand = {
 };
 
 export async function helpInteractionCollector(
-	client: CustomClient,
+	client: Client,
 	interaction: SelectMenuInteraction,
 ): Promise<void> {
 
@@ -48,7 +49,7 @@ export async function helpInteractionCollector(
 	const titleNr = value.substring(value.length - 1);
 	const title = `Page ${titleNr}: ${getArrayElement(titles, Number(titleNr) - 1)}`;
 
-	const description = client.slashCommands
+	const description = handle.slashCommands
 		.filter(c => c.category === getArrayElement(value.split('_'), 1))
 		.sort((c1, c2) => c1.position - c2.position)
 		.map(c => {
