+++ src/events/roleDelete.ts
@@ -1,13 +1,13 @@
 import { Role } from 'discord.js';
 import serverModel from '../models/serverModel';
 import userModel from '../models/userModel';
-import { CustomClient, DiscordEvent, WayOfEarningType } from '../typedef';
+import { DiscordEvent, WayOfEarningType } from '../typedef';
 import { getMapData } from '../utils/helperFunctions';
 
 export const event: DiscordEvent = {
 	name: 'roleDelete',
 	once: false,
-	async execute(client: CustomClient, role: Role) {
+	async execute(client, role: Role) {
 
 		const serverData = await serverModel.findOne(s => s.serverId === role.guild.id);
 		const roles = serverData.shop.filter(shoprole => shoprole.roleId === role.id);
