+++ src/events/shardReady.ts
@@ -1,10 +1,10 @@
 import { ActivityType } from 'discord.js';
-import { CustomClient, DiscordEvent } from '../typedef';
+import { DiscordEvent } from '../typedef';
 
 export const event: DiscordEvent = {
 	name: 'shardReady',
 	once: false,
-	async execute(client: CustomClient) {
+	async execute(client) {
 
 		client.user?.setActivity('/help', { type: ActivityType.Listening });
 	},
