+++ src/utils/updateGuild.ts
@@ -1,14 +1,14 @@
 import { generateId } from 'crystalid';
-import { Guild } from 'discord.js';
+import { Client, Guild } from 'discord.js';
 import { readdirSync, readFileSync, renameSync, writeFileSync } from 'fs';
 import serverModel from '../models/serverModel';
-import { BanList, commonPlantsInfo, CustomClient, DeleteList, materialsInfo, ProxyListType, rarePlantsInfo, ServerSchema, specialPlantsInfo, speciesInfo, uncommonPlantsInfo } from '../typedef';
+import { BanList, commonPlantsInfo, DeleteList, materialsInfo, ProxyListType, rarePlantsInfo, ServerSchema, specialPlantsInfo, speciesInfo, uncommonPlantsInfo } from '../typedef';
 
 /**
  * This creates a new guild if the guild isn't on the ban list, or restores it from the guilds that are to be deleted.
  */
 export async function createGuild(
-	client: CustomClient,
+	client: Client,
 	guild: Guild,
 ): Promise<ServerSchema> {
